Restructure files to eliminate circular dependencies.  Allow a permanent ~/.swift/swiftclock directory.
