+++ swiftbat/__init__.py
@@ -1,4 +1,5 @@
 from __future__ import print_function, division, absolute_import
 
-from .swinfo import *
+from .clockinfo import utcf
 from .swutil import *
+from .swinfo import *
